feat: Handle multiple search results in job processing (#1)
Adds the converted "Plex name" to each job when search results are found, and uses that converted file name as the destination file name for the destination in the conversion phase of job processing. If multiple search results are found, will append a `.N` onto the end of the Plex name to distinguish between conversion results from the same search string.